Move methods into there own classes
* so the nosson generator class isn't so overloaded
* naming should be worked on, but for now this is just a mini cleanup
