+++ NonSucking.Framework.Extension.Generators/NonSucking.Framework.Extension.Generators.csproj
@@ -47,7 +47,7 @@
 
 		<PackageReference Include="System.Composition.AttributedModel" Version="1.0.31" PrivateAssets="all" GeneratePathProperty="true" />
 
-    <PackageReference Include="VaVare" Version="0.16.0.122-alpha" PrivateAssets="all" GeneratePathProperty="true"/>
+    <PackageReference Include="VaVare" Version="0.16.0.122-alpha" PrivateAssets="all" GeneratePathProperty="true" />
     
 		<None Include="$(PkgSystem_Composition_AttributedModel)\lib\netstandard1.0\*.dll" Pack="true" PackagePath="analyzers/dotnet/cs" Visible="false" />
     <None Include="$(PkgVaVare)\lib\netstandard2.0\*.dll" Pack="true" PackagePath="analyzers/dotnet/cs" Visible="false" />
